testcrackpoints added in crackelem_centroid2d
ntestcrackpoint added in parameters module, set to 10
crackelem_centroid2d updated, so that when a crack point cannot be
found using the line intersect algorithm, test crack points on edges
will be used instead. It will loop over all the test points of all
edges, and find the one(or pair) which gives the closest approx. to the
actual crack line